feat: introduce use in viewport
